Improve session component (#58)
* Unrelated: Fight with typescript validation in vscode

I reenabled the builtin typescript extension in vscode but disabled
the validation feature in the workspace. This lets us source format
a document at least.

* Minor: Simplification and type strengthening

I removed double code in the reaction to the server call and
strengthened the type safety by reusing an already existing and used
Session type.